Added coloring to validation warnings. Disabled skip button after task completes
